blacklist playlists not supported by browser media source before initial selection
